fixes tests

+++ tests/flake8_scout_rule_test.py
@@ -22,21 +22,27 @@ def python_dir_with_violations_fixture() -> Generator[str, None, None]:
     with tempfile.TemporaryDirectory() as tmp_dir:
         for f in files:
             copy(f, os.path.join(tmp_dir, os.path.basename(f) + ".py"))
+        os.chdir(tmp_dir)
         yield tmp_dir
+    os.chdir(cwd)
 
 
 @pytest.fixture(autouse=True)
 def flake8_file_fixture() -> Generator[None, None, None]:
     flake8_path = os.path.join(os.getcwd(), ".flake8")
-    assert os.path.isfile(flake8_path) is True
+    if os.path.isfile(flake8_path):
+        temp_flake8_path = os.path.join(os.getcwd(), ".flake8_temp")
+        shutil.move(flake8_path, temp_flake8_path)
+        yield
+        shutil.move(temp_flake8_path, flake8_path)
+    else:
+        yield
 
-    temp_flake8_path = os.path.join(os.getcwd(), ".flake8_temp")
-    shutil.move(flake8_path, temp_flake8_path)
-    yield
-    shutil.move(temp_flake8_path, flake8_path)
 
+def test_black_box_with_path_passed_to_flake8(python_dir_with_violations_fixture: str) -> None:
+    flake8_file = os.path.join(python_dir_with_violations_fixture, ".flake8")
+    assert os.path.isfile(flake8_file) is False
 
-def test_black_box(python_dir_with_violations_fixture: str) -> None:
     command = f"flake8 --format=scout --no-review-prompt {python_dir_with_violations_fixture}"
     result = subprocess.run(
         command,
@@ -45,10 +51,23 @@ def test_black_box(python_dir_with_violations_fixture: str) -> None:
         text=True,
         shell=True,
     )
+
     assert result.returncode == 1  # There should be violations, so it should return 1
     assert result.stderr == ""
-    assert "Found 47 violations" in result.stdout
-    assert "Automatically adding '# noqa: <errors>' annotations" in result.stdout
+    stdout = result.stdout
+
+    assert "Found 47 violations" in stdout
+    assert "Automatically adding '# noqa: <errors>' annotations" in stdout
+    # get the full path since sometimes os.getcwd returns a symlink
+    flake8_file = os.path.realpath(flake8_file)
+    assert f"No flake8 config file found, creating a default one at '{flake8_file}'" in stdout
+    assert os.path.isfile(flake8_file) is True
+    with open(flake8_file, "r") as f:
+        flake8_content = f.read()
+
+    assert "[flake8_scout_rule]\nper_file_violations_tracked = \n\t" in flake8_content
+    assert "\n\tfile1.py: A001{1}, ANN201{1}" in flake8_content
+    assert "\n\tfile2.py: A001{1}, ANN201{1}" in flake8_content
 
     result2 = subprocess.run(
         command,
@@ -58,6 +77,52 @@ def test_black_box(python_dir_with_violations_fixture: str) -> None:
         shell=True,
     )
     assert result2.returncode == 0  # There should be no violations, now they should be noqa'd
+    with open(flake8_file, "r") as f:
+        flake8_content2 = f.read()
+    assert flake8_content2 == flake8_content
+
+
+def test_black_box(python_dir_with_violations_fixture: str) -> None:
+    flake8_file = os.path.join(python_dir_with_violations_fixture, ".flake8")
+    assert os.path.isfile(flake8_file) is False
+
+    command = "flake8 --format=scout --no-review-prompt"
+    result = subprocess.run(
+        command,
+        cwd=python_dir_with_violations_fixture,
+        capture_output=True,
+        text=True,
+        shell=True,
+    )
+
+    assert result.returncode == 1  # There should be violations, so it should return 1
+    assert result.stderr == ""
+    stdout = result.stdout
+
+    assert "Found 47 violations" in stdout
+    assert "Automatically adding '# noqa: <errors>' annotations" in stdout
+    # get the full path since sometimes os.getcwd returns a symlink
+    flake8_file = os.path.realpath(flake8_file)
+    assert f"No flake8 config file found, creating a default one at '{flake8_file}'" in stdout
+    assert os.path.isfile(flake8_file) is True
+    with open(flake8_file, "r") as f:
+        flake8_content = f.read()
+
+    assert "[flake8_scout_rule]\nper_file_violations_tracked = \n\t" in flake8_content
+    assert "\n\tfile1.py: A001{1}, ANN201{1}" in flake8_content
+    assert "\n\tfile2.py: A001{1}, ANN201{1}" in flake8_content
+
+    result2 = subprocess.run(
+        command,
+        cwd=python_dir_with_violations_fixture,
+        capture_output=True,
+        text=True,
+        shell=True,
+    )
+    assert result2.returncode == 0  # There should be no violations, now they should be noqa'd
+    with open(flake8_file, "r") as f:
+        flake8_content2 = f.read()
+    assert flake8_content2 == flake8_content
 
 
 def test_black_box_with_ignore(python_dir_with_violations_fixture: str) -> None:
@@ -199,3 +264,72 @@ def test_flake8_scout_rule_formatter(
         content = f.read()
         assert content.count(noqa_addition) == 1
     assert mock_input.called is True
+
+
+def test_flake8_scout_rule_formatter_no_review_prompt(
+    python_dir_with_violations_fixture: str,
+) -> None:
+    options = Namespace(
+        output_file=None,
+        color=False,
+        tee=False,
+        no_review_prompt=True,
+        no_update_flake8_config=False,
+    )
+    formatter = Flake8ScoutRuleFormatter(options)
+    formatter.start()
+    # Too lazy to add ALL the violations, just add a few
+
+    file1 = python_dir_with_violations_fixture + "/file1.py"
+    file2 = python_dir_with_violations_fixture + "/file2.py"
+    violations = [
+        Violation(
+            code="F841",
+            filename=file1,
+            line_number=4,
+            column_number=5,
+            text="local variable 'nospacebetweenequals' is assigned to but never used",
+            physical_line="    nospacebetweenequalsagain=1\n",
+        ),
+        Violation(
+            code="E225",
+            filename=file1,
+            line_number=4,
+            column_number=25,
+            text="missing whitespace around operator",
+            physical_line="    nospacebetweenequalsagain=1\n",
+        ),
+        Violation(
+            code="F841",
+            filename=file2,
+            line_number=4,
+            column_number=5,
+            text="local variable 'nospacebetweenequals' is assigned to but never used",
+            physical_line="    nospacebetweenequalsagain=1\n",
+        ),
+        Violation(
+            code="E225",
+            filename=file2,
+            line_number=4,
+            column_number=25,
+            text="missing whitespace around operator",
+            physical_line="    nospacebetweenequalsagain=1\n",
+        ),
+    ]
+    for v in violations:
+        formatter.format(v)
+
+    with patch("sys.stdout", new=StringIO()) as captured_stdout:
+        formatter.stop()
+
+    stdout = captured_stdout.getvalue()
+    print(f"Captured stdout:\n{stdout}")
+    assert "Found 4 violations" in stdout
+    noqa_addition = "  # noqa: E225, F841"
+    with open(file1, encoding="UTF-8") as f:
+        content = f.read()
+        assert content.count(noqa_addition) == 1
+
+    with open(file2, encoding="UTF-8") as f:
+        content = f.read()
+        assert content.count(noqa_addition) == 1
